requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pandas>=2.2.0
tqdm>=4.66.0
weaviate-client>=4.5.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-weaviate>=0.0.3
requests>=2.31.0
